Increased Sensitivity (See Dssc)
I played around with the sensitivity to try and find a speed that would feel better. I eventually updated both variables from 400 to 1200.

The camera movement feels smoother, but if you think that it needs to be changed again, please do so. I just think that 1200 seems like a good speed for everyone while we work on our sections.